requirements.txt
# Versi Python yang digunakan
python_version == 3.8

# Modules yang digunakan dalam proyek ini
requests==2.28.1
colorama==0.4.4
dnspython==2.2.1
sublist3r==1.0.7
python-whois==0.7.3
whatweb==0.0.7
wafw00f==2.2.0

# Perintah untuk menginstal pip (biasanya sudah termasuk dalam distribusi Python 3)
pip
